[osx-setup] Host the JDK .pkg internally
I'd like to get Java.Interop building on a continuous integration
system, and if (when) we start building native libraries on OS X we'll
need a way to reliably, *automatically*, download the correct .pkg to
get the JavaVM.framework development headers.

Update the `make osx-setup` target to download an internally-hosted
JavaDeveloper-2013005_dp__11m4609.pkg file, and implicitly download
and extract it when needed.