[Java.Interop, Java.Interop.Export] Add JniValueMarshalerAttribute
Mentioned in b7d50843 was a feature to complete Issue #8:

> Need to permit use of custom attributes on return types, parameter
> types, "inline" into marshal methods.

Let's do that. :-)

The new Java.Interop.JniValueMarshalerAttribute custom attribute can
be applied to:

  * Types: classes, enums, interfaces, structs

This allows:

	[JniValueMarshaler (typeof (MyCustomMarshaler))]
	public class MySpecialClass /* NOT JavaObject! */ {
	}

	public partial class MyCustomMarshaler : JniValueMarshaler<MySpecialClass> {
		// ...
	}

JniRuntime.JniValueManager.GetValueMarshaler(Type) has been updated to
check for the JniValueMarshalerAttribute and, when present, will
return a new instance of the specified JniValueMarshaler instance:

	var marshaler = JniRuntime.CurrentRuntime.ValueManager.GetValueMarshaler (typeof (MySpecialClass));
	// marshaler ISA MyCustomMarshaler

JniValueMarshalerAttribute can also be applied to:

  * Method parameters
  * Method return types

Java.Interop.Export has been updated to check for the
JniValueMarshalerAttribute custom attribute to marshal parameters and
return types, allowing:

	// e.g. code from an existing library
	public class ExistingType {
	}

	public partial class ExistingTypeValueMarshaler : JniValueMarshaler<ExistingType> {
		// ...
	}

	[JavaCallable]
	public static
		[return: JniValueMarshaler (typeof (ExistingTypeValueMarshaler))] ExistingType
		Foo ([JniValueMarshaler (typeof (ExistingTypeValueMarshaler))] ExistingType value)
	{
		return value;
	}

This allows one-off specification or overriding of value marshalers
for method parameter and return types, particularly useful if you want
to marshal a type that you don't control, and thus can't alter to
contain a [JniValueMarshaler] declaration.