[java-interop] `log_gref()` requires explicit newlines.
An oversight in commit c5a6e806 was that Xamarin.Android writes most
messages to `adb logcat`, and thus doesn't require an explicit '\n'.
java-interop, meanwhile, is using fprintf(3), and *does* require an
explicit newline, which was forgotten in the migration.

Oops.

Add "\n" to log_gref() use so that logs are more readable.