[Java.Interop] Change com.xamarin.android to com.xamarin.java_interop
[@dellis1972 noted][0]:

> Just a thought, currently Java.Interop is not dependent on Android
> as far as I can tell. do we really want to use an "android" package
> name here. perhaps com.xamarin.javainterop might be more generic.

This is an excellent point, and I'm not sure why I didn't think of it
in the first place, as desktop use/support was one of the original
design requirements and naming guidelines.

Rename the `com.xamarin.android` package prefix to
`com.xamarin.java_interop`.

(Why `java_interop`? Because it goes nicely with "java-interop" and
"Java.Interop", and a nested `java` package such as
`com.xamarin.java.interop` seems counter to Java convention.)

[0]: https://github.com/xamarin/Java.Interop/commit/dd099847f2a5edefbb16d1658b32664259f3414a#commitcomment-14707370