[Java.Interop] Change JniObjectReference.ToString() format.
Change JniObjectReference.ToString() from the fairly verbose and
context-free:

	JniObjectReference(Handle=0x8115ef58, Type=Global)

to the terse and context-required:

	0x8115ef58/G

Why? Consistency and `grep`. The latter is the format used in the JNI
handle logging, e.g.

	+g+ grefc 7 gwrefc 0 obj-handle 0x7e761450/L -> new-handle 0x8115ef58/G from thread 'TestRunnerThread'(6)

Or, in context:

	+g+ grefc 5 gwrefc 0 obj-handle 0x7e761448/L -> new-handle 0x8115ef58/G from thread 'TestRunnerThread'(6)
	...
	Created PeerReference=0x8115ef58/G IdentityHashCode=0x6e70c7 Instance=0x18bbccbc Instance.Type=Java.Interop.JavaBooleanArray, Java.Type=[Z

Since the handles now use the same format, you can copy & paste that
string and find both occurrences. Previously, the log file would have:

	+g+ grefc 5 gwrefc 0 obj-handle 0x7e761448/L -> new-handle 0x8115ef58/G from thread 'TestRunnerThread'(6)
	...
	Created PeerReference=JniObjectReference(Handle=0x8115ef58, Type=Global) IdentityHashCode=0x6e70c7 Instance=0x18bbccbc Instance.Type=Java.Interop.JavaBooleanArray, Java.Type=[Z

Searching for the pointer value works, but not pointer + type.