[Java.Interop] JniRuntime.JniValueManager method renames.
Rename JniRuntime.JniValueManager.PeekObject() to PeekPeer().
Rationale: consistent terminology (99b16fb0). What's an "Object"?

Rename JniRuntime.JniValueManager.GetSurfacedObjects() to
GetSurfacedPeers().
Rationale: consistent terminology (99b16fb0). What's an "Object"?

Change the return type of
JniRuntime.JniValueManager.GetSurfacedPeers() from
List<WeakReference<IJavaPeerable>> to
List<JniSurfacedPeerInfo>>:

	- public abstract List<WeakReference<IJavaPeerable>>  GetSurfacedObjects ();
	+ public abstract List<JniSurfacedPeerInfo>           GetSurfacedPeers ();

Rationale: While trying to investigate object reference lifetimes in
Xamarin.Android, it is not unusual for
Java.Interop.Runtime.GetSurfacedObjects() -- which returns a
List<WeakReference> -- to contain a WeakReference which has been
collected -- i.e. the WeakReference is non-null while
WeakReference.Target is null.

Meaning I *know* there's an identityHashCode <=> peer mapping, but
there's no way -- in the *public* API -- to know what the
identityHashCode value *is*. (The identityHashCode value is desired
because that's also printed in the GREF log, so with the
identityHashCode value I could at least determine what *type* it was.)

Which makes it very hard to reason about anything without adding some
Console.WriteLine()s into the middle of things, to obtain information
that was "lost" because it was never provided.

The fix? Provide the identityHashCode value:

	class JniSurfacedPeerInfo {
		public  int                             JniIdentityHashCode     {get; private set;}
		public  WeakReference<IJavaPeerable>    SurfacedPeer            {get; private set;}
	}

JniSurfacedPeerInfo.JniIdentityHashCode is the original
identityHashCode value for the instance held by
JniSurfacedPeerInfo.SurfacedPeer, *even if* the peer has been
collected.

JniSurfacedPeerInfo is a class so that it can be optionally subclassed.