[Java.Interop] Rename ManagedPeer.runConstructor() to construct()
Commit dd09984 decided to drop the Xamarin.Android pattern:

	// Xamarin.Android constructor style
	public JavaPeerType(Args... args) {
		super (args...);
		if (getClass() == JavaPeerType.class)
			TypeManager.Activate (
					AssemblyQualifiedTypeName,
					ConstructorSignature,
					this,
					new java.lang.Object[]{ args }
			);
	}

And "move" the getClass() check into a helper method:

	// Java.Interop constructor style:
	public JavaPeerType(Args... args) {
		super (args...);
		ManagedPeer.runConstructor (
				JavaPeerType.class,
				this,
				AssemblyQualifiedTypeName,
				ConstructorSignature,
				args...
		);
	}

Unspecified was *why* this would be a good thing; the primary thought
was that this might reduce .class size (unverified!), as there
wouldn't need to be the `if` check.

Upon further reflection, while the dd09984 approach looks cleaner, it
is *less* flexible.

One of the future features to work toward is FullAOT support
(326a1b2b, 1eef4653), and one of the problems there is...activation
constructors. :-)

What I'd like to explore doing is updating jnimarshalmethod-gen.exe to
emit constructors. This mixes up the ManagedPeer.runConstructor()
paradigm because in such an environment we'd instead want `native`
methods in the Java Callable Warpper which the constructor would
invoke:

	// Java
	class MyCallableWrapper {
		public MyCallableWrapper () {
			// What should go here?
		}
		native void n_ctor ();
	}

The obvious thing to do is...a Xamarin.Android-style type check!

	class MyCallableWrapper {
		public MyCallableWrapper () {
			if (MyCallableWrapper.class == getClass() {
				n_ctor ();
			}
		}
		native void n_ctor ();
	}

The dd09984-style ManagedPeer.runConstructor() idiom can't support
that, not when the "constructor" native methods are per-class.

Thus, there's no actual need for doing the getClass() check within
ManagedPeer.runConstructor().

Remove this overload:

	ManagedPeer.runConstructor(Class, Object, String, String, Object[])

Rename ManagedPeer.runConstructor(Object, String, String, Object[]) to
ManagedPeer.construct(), and change the existing Java types to use a
Xamarin.Android-style type check.