[Java.Interop] Use exception filters with JniTransition
Exception handling [0] in .NET is based on many of the concepts of,
and on Windows *is*, WinAPI Structured Exception Handling (SEH) [1].

When an exception is thrown:

 1. If a debugger is attached, it will be given an opportunity, a
    "first chance exception", to "break" before the exception is
    thrown.

 2. The call stack is *scanned* to see if the exception will be
    handled.

 2.a. If a debugger is attached and the exception is not handled,
      then the debugger will "break".

 3. When execution continues (or there is no debugger...), the call
    stack will be unwound. If a frame contains `catch` blocks, they
    may be processed (and unwinding will stop). If a frame contains
    `finally` blocks, they will be executed.

For example, if a debugger is attached this will cause the debugger to
break on an unhandled exception:

	void Foo ()
	{
		throw new Exception ("Die!");
	}

while this will *not* break on an unhandled exception, as the
exception is handled by the `catch` block:

	void Bar ()
	{
		try {
			Foo ();
		}
		catch (Exception e) {
		}
	}

This "two pass" exception handling strategy (steps (2) and (3)) is
unique to Windows and .NET, and is not a construct in JNI.

This raises an interesting conundrum: what happens for "interleaved"
call stacks, in which both Java and managed frames are present?

	Managed > Java > Managed (...)

Whenever Java code calls into managed code, the first managed frame
uses JniTransition (26a93217):

	void JavaToManagedFrame (IntPtr jnienv, IntPtr context, ...)
	{
		var t = new JnITransition (jnienv);
		try {
			...
		}
		catch (Exception e) {
			t.SetPendingException (e);
		}
		finally {
			t.Dispose ();
		}
	}

Which means *every* Java to managed transition frame catches all
exceptions!

THERE ARE NO UNHANDLED EXCEPTIONS.

For normal use, this is fine. When debugging the app, this is NOT
fine, because the debugger thinks that nearly all exceptions are
unhandled. The only unhandled exceptions will be those in which there
is no Java "parent" stack frame, e.g. new Threads, and even then the
debugger will "break" at the "bottommost" stack frame, FAR AWAY from
where the exception was originally thrown.

This makes for a sucky developer experience.

To better support this, use C# exception filters [2] so that instead
of always catching all exceptions, we only catch them when we want to:

	void JavaToManagedFrame (IntPtr jnienv, IntPtr context, ...)
	{
		var t = new JnITransition (jnienv);
		try {
			...
		}
		catch (Exception e) when (JniEnvironment.Runtime.ExceptionShouldTransitionToJni (e)) {
			t.SetPendingException (e);
		}
		finally {
			t.Dispose ();
		}
	}

We do this by introducing a new
JniRuntime.ExceptionShouldTransitionToJni() method. If it returns
`true`, then the catch block will be executed. If it returns `false`,
then the catch block will be *ignored*...during *both* passes.

THERE BE DRAGONS: The problem with ignoring the `catch` block during
*both* passes is that it means JNI is *skipped* from stack unwinding!
Consequently, if:

 1. A managed exception is thrown, and
 2. A debugger is attached

then the managed runtime will take responsibility for unwinding *all*
frames, and *this will include the Java frames*. This in turn
introduces the "wonderful" scenario of *skipping over* Java frames
during exception handling cleanup.

No good can come from this. It can corrupt JVM state and result in
process aborts.

BE EXTREMELY CAREFUL returning `false` from
JniRuntime.ExceptionShouldTransitionToJni(). It should NOT be the
norm; use it ONLY while debugging, and expect the process to possibly
crash soon afterward.

Update hand-written JniTransition code to use
JniRuntime.ExceptionShouldTransitionToJni(), and update
MarshalMemberBuilder to likewise use it for generated marshal methods.

[0]: https://msdn.microsoft.com/en-us/library/x85tt0dd.aspx
[1]: https://www.microsoft.com/msj/0197/exception/exception.aspx
[2]: http://www.thomaslevesque.com/2015/06/21/exception-filters-in-c-6/