[Java.Interop] Invalidate JniEnvironment when JniRuntime is disposed
(One part "hygiene", one part "expectations", and one part "how do we
test JniRuntime.Dispose() behavior?"...)

You have a JVM:

	$ make shell
	csharp> using Java.Interop;
	csharp> var vm = new Java.InteropTests.TestJVM (new string[0]);

It has a JNIEnv* value:

	csharp> var envp = JniEnvironment.EnvironmentPointer;
	csharp> envp;
	-2102644448

Through it, you can do things with JNI:

	csharp> var t = new JniType ("java/lang/Object");

JVM's can also be destroyed:

	csharp> vm.Dispose();

Thus, a "philosophical" question: what happens if you try to use the
JVM after it's been destroyed?

Is "horrifically crash the process" your answer?

	csharp> new JniType ("java/lang/Object");
	Stacktrace:

	  at <unknown> <0xffffffff>
	  at (wrapper managed-to-native) Java.Interop.NativeMethods.java_interop_jnienv_find_class (intptr,intptr&,string) <IL 0x0003d, 0xffffffff>
	  at Java.Interop.JniEnvironment/Types.FindClass (string) [0x0003c] in /Users/jon/Dropbox/Developer/xamarin/Java.Interop/src/Java.Interop/Java.Interop/JniEnvironment.Types.cs:41
	  at Java.Interop.JniType..ctor (string) [0x00008] in /Users/jon/Dropbox/Developer/xamarin/Java.Interop/src/Java.Interop/Java.Interop/JniType.cs:30
	...

The process aborts. You have no chance to survive. Make your time [0].

That's...less than desirable?

Integrate JniEnvironment and JniRuntime.Dispose() so that when
the JniRuntime is disposed, all "associated JniEnvironments" are
invalidated. That turns the above SIGSEGV into a
NotSupportedException:

	csharp> new JniType("java/lang/Object");
	System.NotSupportedException: JNI Environment Information has been invalidated on this thread.
	  at Java.Interop.JniEnvironment.get_CurrentInfo () [0x00029] in /Users/jon/Dropbox/Developer/Java.Interop/src/Java.Interop/Java.Interop/JniEnvironment.cs:20
	  at Java.Interop.JniEnvironment+Types.FindClass (System.String classname) [0x0002d] in /Users/jon/Dropbox/Developer/Java.Interop/src/Java.Interop/Java.Interop/JniEnvironment.Types.cs:38
	  at Java.Interop.JniType..ctor (System.String classname) [0x00008] in /Users/jon/Dropbox/Developer/Java.Interop/src/Java.Interop/Java.Interop/JniType.cs:30

Of course, we need to be able to *test* this, and to do *that* it
would be quite handy to be able to (sanely?) have more than one
JniRuntime exist in a process at a time.

...yet JNI_CreateJavaVM() can only ever create one runtime, so how can
we have more than one JniRuntime?

Aliasing!

It doesn't seem *that* unreasonable for multiple JniRuntime instances
to share the same JavaVM* pointer, at least for unit-test purposes, so
fix any obvious breakage that results when you do so:

	var o = new JniRuntime.CreationOptions {
		InvocationPointer = JniRuntime.Current.InvocationPointer,
		// ...
	};

This is conceptually "iffy" -- see 4f5a8e94 and the interraction of
the GC and JniRuntime instances -- but not *impossible*, and allowing
such aliasing allows us to create a new JniRuntime instance on a
dedicated thread, Dispose() it, and ensure that accessing
JniEnvironment.Runtime throws appropriately.

Which brings us to better supporting this "conceptually 'iffy'"
construct: The JniRuntime constructor should explicitly create the
JNIEnv* for the current thread, instead of the previous approach of
"implicitly" doing this when creating a JniType instance.

Why? Because:

 1. We don't want the JniRuntime constructor to always set
    JniRuntime.CurrentRuntime, so

 2. The "implicit JNIEnv* attach" will result in using the *current*
    JniRuntime.CurrentRuntime value, not the one we're constructing!

Implicit behavior BAD. Explicit behavior GOOD (or better?).

We also add a new JniRuntime.CreationOptions.JniVersion property so
that the JniVersion we attach to the JVM with can be controlled.

[0]: https://www.youtube.com/watch?v=8fvTxv46ano