[build] Improve JDK path probing support
The problem with commit 20c97cee, which attempted to host the JDK .pkg
internally, is threefold:

 1. It's inherently OS specific, with no framework for supporting
    operating systems other than OS X; and

 2. It's an *ancient* .pkg (for Java 6!), and the world has moved on.

 3. The internal URL won't be publicly accessible for much longer.

As it happens™, monodroid has many of the same problems: it needs to
support a variety of JDK locations on a variety of operating systems.

"Munge" monodroid's JDK path probing infrastructure and alter for use
within Java.Interop.

This change creates a new bin\Build$(Configuration)\JdkHeaders.props
file, which is an MSBuild property file which contains
@(JdkIncludePath) values for all directories which contain native
headers which should be provided to the compiler.

This change also creates a new make(1) variable,
$(JI_JDK_INCLUDE_PATHS), which likewise contains all the directories
which contain JDK headers which need to be provided to the compiler

At present only OS X and Linux are supported.