[build] Don't build the Android* projects.
The Android* projects are somewhat vestigial, in that
Xamarin.Android 6.1's Mono.Andriod.dll now uses Java.Interop.dll, so
the need for "intra-Java.Interop Xamarin.Android support" is gone.
(That's also the purpose of the XAIntegrationDebug and
XAIntegrationRelease configurations).

Additionally, we don't want to require Xamarin.Android to be installed
on Jenkins. :-)

Keep the projects, but remove them from the Debug and Release
configurations so that they won't be built.