[generator] Emit GeneratedFiles.projitems, not GeneratedFiles.csproj
Commit a32474f0 had the right idea, but the wrong extension. `.csproj`
is a C# *Project* file, generally meaning that it has useful OS
support, e.g. double-clicking on a .csproj file will open the file
into an IDE/editor/etc.

The problem is that the generated file, while in MSBuild XML format,
isn't a *project* file. It can't be used, by itself, to build
*anything*. It's just a list of files (and possibly defines).

Instead of emitting a `GeneratedFiles.csproj` file, emit a
`GeneratedFiles.projitems` file, to make it clearer that the file
contains *Project Items*. (This is also the extension used by
Shared Projects for the files contained within the Shared Project.)